feat(zip): Refactor code structure to support zip file parsing.
