+++ src/main/java/com/synopsys/integration/create/apigen/parser/ApiParser.java
@@ -23,141 +23,9 @@
 package com.synopsys.integration.create.apigen.parser;
 
 import java.io.File;
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
-import com.google.gson.Gson;
-import com.synopsys.integration.create.apigen.Application;
-import com.synopsys.integration.create.apigen.data.MediaTypes;
-import com.synopsys.integration.create.apigen.data.MissingFieldsAndLinks;
-import com.synopsys.integration.create.apigen.data.NameAndPathManager;
-import com.synopsys.integration.create.apigen.data.TypeTranslator;
-import com.synopsys.integration.create.apigen.data.UtilStrings;
-import com.synopsys.integration.create.apigen.model.DefinitionParseParameters;
-import com.synopsys.integration.create.apigen.model.FieldDefinition;
 import com.synopsys.integration.create.apigen.model.ParsedApiData;
-import com.synopsys.integration.create.apigen.model.RawFieldDefinition;
-import com.synopsys.integration.create.apigen.model.RequestDefinition;
-import com.synopsys.integration.create.apigen.model.ResponseDefinition;
 
-public class ApiParser {
-    private final MediaTypes mediaTypes;
-    private final Gson gson;
-    private final TypeTranslator typeTranslator;
-    private final NameAndPathManager nameAndPathManager;
-    private final MissingFieldsAndLinks missingFieldsAndLinks;
-
-    public ApiParser(final MediaTypes mediaTypes, final Gson gson, final TypeTranslator typeTranslator, final NameAndPathManager nameAndPathManager, final MissingFieldsAndLinks missingFieldsAndLinks) {
-        this.mediaTypes = mediaTypes;
-        this.gson = gson;
-        this.typeTranslator = typeTranslator;
-        this.nameAndPathManager = nameAndPathManager;
-        this.missingFieldsAndLinks = missingFieldsAndLinks;
-    }
-
-    public ParsedApiData parseApi(final File specificationRootDirectory) {
-        final File endpointsPath = new File(specificationRootDirectory, "endpoints");
-        final File apiPath = new File(endpointsPath, UtilStrings.API);
-
-        List<RequestDefinition> requestDefinitions = new LinkedList<>();
-        List<ResponseDefinition> responseDefinitions = new LinkedList<>();
-        parseApi(apiPath, apiPath.getAbsolutePath().length() + 1, requestDefinitions, responseDefinitions);
-        return new ParsedApiData(requestDefinitions, responseDefinitions);
-    }
-
-    private void parseApi(final File parent, final int prefixLength, List<RequestDefinition> requestDefinitions, List<ResponseDefinition> responseDefinitions) {
-        final List<File> children = Arrays.stream(parent.listFiles())
-                                        .filter(file -> !file.getName().equals("notifications"))
-                                        .sorted()
-                                        .collect(Collectors.toList());
-
-        // If child file of parent is response specification data, parse the file, otherwise recurse and parse the child's children
-        for (final File child : children) {
-            String fileName = child.getName();
-            boolean requestOrResponseFile = fileName.equals(UtilStrings.REQUEST_SPECIFICATION_JSON) || fileName.equals(UtilStrings.RESPONSE_SPECIFICATION_JSON);
-            if (requestOrResponseFile && parent.getAbsolutePath().contains(Application.RESPONSE_TOKEN)) {
-                final String relativePath = child.getAbsolutePath().substring(prefixLength);
-                final String mediaType = mediaTypes.getLongName(child.getParentFile().getName());
-                if (fileName.equals(UtilStrings.REQUEST_SPECIFICATION_JSON)) {
-                    requestDefinitions.add(new RequestDefinition(relativePath, mediaType));
-                } else if (fileName.equals(UtilStrings.RESPONSE_SPECIFICATION_JSON)) {
-                    final NameParser nameParser = new NameParser(nameAndPathManager);
-                    final String responseName = nameParser.getResponseName(relativePath);
-                    final String responseMediaType = mediaTypes.getLongName(child.getParentFile().getName());
-                    final boolean doesHaveMultipleResults = computeIfHasMultipleResults(child);
-                    responseDefinitions.add(new ResponseDefinition(relativePath, responseName, responseMediaType, doesHaveMultipleResults));
-                }
-            } else if (child.isDirectory()) {
-                parseApi(child, prefixLength, requestDefinitions, responseDefinitions);
-            }
-        }
-    }
-
-    private boolean computeIfHasMultipleResults(final File file) {
-        final File parentOfArrayResponse = getParentOfArrayResponse(file);
-        if (parentOfArrayResponse == null) {
-            return false;
-        }
-        return containsArrayResponse(parentOfArrayResponse, false);
-    }
-
-    private File getParentOfArrayResponse(final File file) {
-        File current = file;
-        while (!current.getName().equals(UtilStrings.GET)) {
-            current = current.getParentFile();
-        }
-        final File getFile = current;
-        current = getFile.getParentFile();
-        if (current.getParentFile() != null && !current.getParentFile().getName().equals(UtilStrings.API)) {
-            return current.getParentFile();
-        } else {
-            return null;
-        }
-    }
-
-    private boolean containsArrayResponse(final File file, final boolean isGetFileSubDirectory) {
-        final File[] children = file.listFiles();
-        if (children == null) {
-            return false;
-        }
-        for (final File child : children) {
-            if (child.getName().equals(UtilStrings.RESPONSE_SPECIFICATION_JSON)) {
-                final ResponseDefinition potentialArrayResponse = buildDummyResponseDefinitionFromFile(child);
-                if (!ResponseTypeIdentifier.getResponseType(potentialArrayResponse).equals(ResponseType.STANDARD)) {
-                    return true;
-                }
-            }
-            if (child.getName().equals(UtilStrings.GET) || (isGetFileSubDirectory && !child.getName().equals(UtilStrings.REQUEST_SPECIFICATION_JSON))) {
-                return containsArrayResponse(child, true);
-            }
-        }
-        return false;
-    }
-
-    public ResponseDefinition extractResponseFromSubfieldsOfItems(ResponseDefinition response) {
-        FieldDefinition items = null;
-        for (FieldDefinition field : response.getFields()) {
-            if (field.getPath().equals(UtilStrings.ITEMS)) {
-                items = field;
-            }
-        }
-        ResponseDefinition newResponse = new ResponseDefinition(response.getResponseSpecificationPath(), response.getName(), response.getMediaType(), response.hasMultipleResults());
-        newResponse.addFields(items.getSubFields());
-
-        return newResponse;
-    }
-
-    private ResponseDefinition buildDummyResponseDefinitionFromFile(final File file) {
-        final DefinitionParser definitionParser = new DefinitionParser(gson, file);
-        final Set<RawFieldDefinition> rawFieldDefinitions = definitionParser.getDefinitions(DefinitionParseParameters.RAW_FIELD_PARAMETERS);
-        final FieldDefinitionProcessor fieldDefinitionProcessor = new FieldDefinitionProcessor(typeTranslator, nameAndPathManager, missingFieldsAndLinks);
-        final Set<FieldDefinition> fieldDefinitions = fieldDefinitionProcessor.parseFieldDefinitions("", rawFieldDefinitions);
-        final ResponseDefinition response = new ResponseDefinition("", "", "", false);
-        response.addFields(fieldDefinitions);
-        return response;
-    }
+public interface ApiParser {
+    ParsedApiData parseApi(File target);
 }
-
