+++ src/main/java/com/synopsys/integration/create/apigen/parser/DirectoryWalker.java
@@ -24,85 +24,51 @@ package com.synopsys.integration.create.apigen.parser;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.Gson;
 import com.synopsys.integration.create.apigen.GeneratorRunner;
-import com.synopsys.integration.create.apigen.data.MediaTypes;
-import com.synopsys.integration.create.apigen.data.MissingFieldsAndLinks;
-import com.synopsys.integration.create.apigen.data.NameAndPathManager;
-import com.synopsys.integration.create.apigen.data.TypeTranslator;
-import com.synopsys.integration.create.apigen.model.DefinitionParseParameters;
-import com.synopsys.integration.create.apigen.model.LinkDefinition;
 import com.synopsys.integration.create.apigen.model.ParsedApiData;
-import com.synopsys.integration.create.apigen.model.RawFieldDefinition;
 import com.synopsys.integration.create.apigen.model.ResponseDefinition;
 
 public class DirectoryWalker {
-    private final File rootDirectoryFile;
+    private static final Logger logger = LoggerFactory.getLogger(GeneratorRunner.class);
     private final Gson gson;
-    private final MediaTypes mediaTypes;
-    private final TypeTranslator typeTranslator;
-    private final NameAndPathManager nameAndPathManager;
-    private final MissingFieldsAndLinks missingFieldsAndLinks;
-    private final Logger logger = LoggerFactory.getLogger(GeneratorRunner.class);
+    private final ApiParser apiParser;
 
-    public DirectoryWalker(final File rootDirectory, final Gson gson, final MediaTypes mediaTypes, final TypeTranslator typeTranslator, final NameAndPathManager nameAndPathManager, final MissingFieldsAndLinks missingFieldsAndLinks) {
-        this.rootDirectoryFile = rootDirectory;
+    public DirectoryWalker(final Gson gson, ApiParser apiParser) {
         this.gson = gson;
-        this.mediaTypes = mediaTypes;
-        this.typeTranslator = typeTranslator;
-        this.missingFieldsAndLinks = missingFieldsAndLinks;
-        this.nameAndPathManager = nameAndPathManager;
+        this.apiParser = apiParser;
     }
 
-    public ParsedApiData parseDirectoryForResponses(final boolean showOutput, final boolean controlRun) throws IOException {
-        final ApiParser apiParser = new ApiParser(mediaTypes, gson, typeTranslator, nameAndPathManager, missingFieldsAndLinks);
-        final FieldDefinitionProcessor processor = new FieldDefinitionProcessor(typeTranslator, nameAndPathManager, missingFieldsAndLinks);
-        boolean actuallyShowOutput = showOutput;
-
+    public ParsedApiData parseDirectoryForResponses(boolean showOutput, boolean controlRun, File target) throws IOException {
         // Get response-specification.json files from directory
-        ParsedApiData parsedApiData = apiParser.parseApi(rootDirectoryFile);
+        ParsedApiData parsedApiData = apiParser.parseApi(target);
         final List<ResponseDefinition> responseDefinitions = parsedApiData.getResponseDefinitions();
-        final List<ResponseDefinition> finalResponseDefinitions = new ArrayList<>();
 
-        // For each response file, parse the JSON for FieldDefinition objects
-        for (final ResponseDefinition response : responseDefinitions) {
-            final String absolutePath = rootDirectoryFile.getAbsolutePath() + "/endpoints/api/" + response.getResponseSpecificationPath();
-            final File responseSpecificationFile = new File(absolutePath);
-            final DefinitionParser definitionParser = new DefinitionParser(gson, responseSpecificationFile);
-
-            final Set<RawFieldDefinition> fields = definitionParser.getDefinitions(DefinitionParseParameters.RAW_FIELD_PARAMETERS);
-            final Set<LinkDefinition> links = definitionParser.getDefinitions(DefinitionParseParameters.LINK_PARAMETERS);
-            response.addFields(processor.parseFieldDefinitions(response.getName(), fields));
-            response.addLinks(links);
-
-            // Filter out 'Array' responses and extract data from responses where data is subfield of "items" field
-            ResponseType responseType = ResponseTypeIdentifier.getResponseType(response);
-            if (responseType.equals(ResponseType.ARRAY)) {
-                actuallyShowOutput = false;
-            } else if (responseType.equals(ResponseType.DATA_IS_SUBFIELD_OF_ITEMS)) {
-                finalResponseDefinitions.add(apiParser.extractResponseFromSubfieldsOfItems(response));
-            } else {
-                finalResponseDefinitions.add(response);
-            }
-
-            if (actuallyShowOutput) {
-                logger.info("***********************\n" + gson.toJson(response));
-            }
-            actuallyShowOutput = showOutput;
-
+        // log output
+        if (showOutput) {
+            logResponses(responseDefinitions);
         }
         // Write output of FieldsParser to test data file
         if (controlRun) {
-            FieldsParserTestDataCollector.writeControlData(gson, finalResponseDefinitions);
+            FieldsParserTestDataCollector.writeControlData(gson, responseDefinitions);
         }
-        return new ParsedApiData(parsedApiData.getRequestDefinitions(), finalResponseDefinitions);
+        return parsedApiData;
+    }
+
+    private void logResponses(Collection<ResponseDefinition> responseDefinitions) {
+        for (final ResponseDefinition response : responseDefinitions) {
+            // Filter out 'Array' responses
+            ResponseType responseType = ResponseTypeIdentifier.getResponseType(response);
+            if (!responseType.equals(ResponseType.ARRAY)) {
+                logger.info("***********************\n" + gson.toJson(response));
+            }
+        }
     }
 
 }
