+++ src/main/java/com/synopsys/integration/create/apigen/parser/FieldDefinitionProcessor.java
@@ -22,10 +22,8 @@
  */
 package com.synopsys.integration.create.apigen.parser;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
