+++ src/test/java/com/synopsys/integration/create/apigen/parser/DirectoryWalkerTest.java
@@ -20,6 +20,7 @@ import com.synopsys.integration.create.apigen.data.MissingFieldsAndLinks;
 import com.synopsys.integration.create.apigen.data.NameAndPathManager;
 import com.synopsys.integration.create.apigen.data.TypeTranslator;
 import com.synopsys.integration.create.apigen.model.ParsedApiData;
+import com.synopsys.integration.create.apigen.parser.file.DirectoryPathParser;
 
 public class DirectoryWalkerTest {
     private static final String API_SPEC_PATH = "api-specification/2019.12.0";
@@ -28,13 +29,19 @@ public class DirectoryWalkerTest {
     private final com.synopsys.integration.create.apigen.parser.DirectoryWalker directoryWalker;
 
     public DirectoryWalkerTest() throws URISyntaxException {
-        this.directoryWalker = new DirectoryWalker(new File(rootDirectory.toURI()), gson, new MediaTypes(), new TypeTranslator(), new NameAndPathManager(), new MissingFieldsAndLinks());
+        MediaTypes mediaTypes = new MediaTypes();
+        TypeTranslator typeTranslator = new TypeTranslator();
+        NameAndPathManager nameAndPathManager = new NameAndPathManager();
+        MissingFieldsAndLinks missingFieldsAndLinks = new MissingFieldsAndLinks();
+        final FieldDefinitionProcessor processor = new FieldDefinitionProcessor(typeTranslator, nameAndPathManager, missingFieldsAndLinks);
+        final DirectoryPathParser apiParser = new DirectoryPathParser(mediaTypes, gson, typeTranslator, nameAndPathManager, missingFieldsAndLinks, processor);
+        this.directoryWalker = new DirectoryWalker(gson, apiParser);
     }
 
     @Test
     public void test() throws IOException, URISyntaxException {
         final File testFile = new File("./build/FieldsParserTestTestingData.txt");
-        ParsedApiData apiData = directoryWalker.parseDirectoryForResponses(false, false);
+        ParsedApiData apiData = directoryWalker.parseDirectoryForResponses(false, false, new File(rootDirectory.toURI()));
         FieldsParserTestDataCollector.writeTestData(gson, apiData.getResponseDefinitions(), testFile);
         final File controlFile = new File(DirectoryWalkerTest.class.getClassLoader().getResource("FieldsParserTestControlData.txt").toURI());
